Move all measurement objects into one file (avoiding circular dependnecies)
