+++ packages/components/src/internal/util/measurement.ts
@@ -1,4 +1,115 @@
-import { BASE_UNITS, MeasurementUnit } from './models';
+import { immerable } from 'immer';
+
+export enum BASE_UNITS {
+    COUNT = 'Count',
+    MASS = 'Mass',
+    VOLUME = 'Volume',
+}
+
+export class UnitModel {
+    [immerable] = true;
+
+    readonly value: number;
+    readonly unitStr?: string; //Originating unit string, may not be a typical MeasurementUnit
+    readonly unit?: MeasurementUnit;
+
+    constructor(value: number, unitStr: string) {
+        const unit = MEASUREMENT_UNITS[unitStr?.toLowerCase()] || null;
+        Object.assign(this, {value, unitStr, unit});
+    }
+
+    isSupportedUnitType(): boolean {
+        return this.unit != null;
+    }
+
+    canConvert(newUnitStr: string): boolean {
+        if (!newUnitStr) {
+            return this.unit == null;
+        }
+
+        const newUnit: MeasurementUnit = MEASUREMENT_UNITS[newUnitStr.toLowerCase()];
+        return newUnit?.baseUnit == this.unit?.baseUnit;
+    }
+
+    as(newUnitStr: string): UnitModel {
+        if (!this.canConvert(newUnitStr)) {
+            throw new Error('Cannot convert to "' + newUnitStr + '"');
+        }
+
+        const newUnit = MEASUREMENT_UNITS[newUnitStr?.toLowerCase()];
+        if (!newUnit) {
+            throw new Error('Unit type not supported "' + newUnitStr + '"');
+        }
+
+        if (this.value == null) {
+            return new UnitModel(0, newUnit.label.toLowerCase());
+        }
+
+        const newValue = this.value * (this.unit.ratio / newUnit.ratio);
+        return new UnitModel(parseFloat(newValue.toFixed(6)), newUnit.label.toLowerCase());
+    }
+
+    add(deltaValue: number, deltaUnitStr?: string) {
+        let deltaUnit: MeasurementUnit = this.unit;
+        if (deltaUnitStr) {
+            if (!this.canConvert(deltaUnitStr)) {
+                throw new Error('Cannot add "' + deltaUnitStr + '" to "' + this.unit?.label + '"');
+            }
+
+            deltaUnit = MEASUREMENT_UNITS[deltaUnitStr.toLowerCase()];
+            if (!deltaUnit) {
+                throw new Error('Unit type not supported "' + deltaUnitStr + '"');
+            }
+        }
+
+        if (!this.unit) {
+            return new UnitModel(this.value + deltaValue, this.unit?.label?.toLowerCase());
+        }
+
+        const newValue = this.value + deltaValue * (deltaUnit.ratio / this.unit.ratio);
+        return new UnitModel(newValue, this.unit.label.toLowerCase());
+    }
+
+    compareTo(other: UnitModel) {
+        const unitStr = this.unit?.label;
+        const otherUnitStr = other.unit?.label;
+        if (!areUnitsCompatible(unitStr, otherUnitStr)) {
+            return undefined;
+        }
+
+        if (this.value == null && other.value != null) {
+            return undefined;
+        }
+        if (this.value != null && other.value == null) {
+            return undefined;
+        }
+
+        if (unitStr == otherUnitStr) {
+            return this.value - other.value;
+        }
+
+        return this.value - other.as(unitStr).value;
+    }
+
+    toString(): string {
+        return this.value + (this.unit ? (' ' + this.unit.label) : '');
+    }
+
+    toDisplayString(): string {
+        return (this.value ? this.value.toLocaleString() : 0) + (this.unit ? (' ' + this.unit.label) : '');
+    }
+
+}
+
+export interface MeasurementUnit {
+    baseUnit: BASE_UNITS;
+    // Number of decimal places allowed when unit is displayed
+    displayPrecision: number;
+    label: string;
+    longLabelPlural: string;
+    longLabelSingular: string;
+    ratio: number;
+}
 
 export const MEASUREMENT_UNITS: { [key: string]: MeasurementUnit } = {
     g: {
@@ -64,34 +175,52 @@ export const MEASUREMENT_UNITS: { [key: string]: MeasurementUnit } = {
  * @param unitBStr
  */
 export function areUnitsCompatible(unitAStr: string, unitBStr: string) {
-    if (unitAStr == unitBStr) return true;
-    if (!unitAStr && !unitBStr) return true;
-    if (unitAStr && !unitBStr) return false;
-    if (!unitAStr && unitBStr) return false;
+    if (unitAStr == unitBStr) {
+        return true;
+    }
+    if (!unitAStr && !unitBStr) {
+        return true;
+    }
+    if (unitAStr && !unitBStr) {
+        return false;
+    }
+    if (!unitAStr && unitBStr) {
+        return false;
+    }
     const unitA: MeasurementUnit = MEASUREMENT_UNITS[unitAStr.toLowerCase()];
     const unitB: MeasurementUnit = MEASUREMENT_UNITS[unitBStr.toLowerCase()];
-    if (!unitA || !unitB) return false;
+    if (!unitA || !unitB) {
+        return false;
+    }
     return unitA.baseUnit == unitB.baseUnit;
 }
 
 export function getMetricUnitOptions(): any[] {
     const options = [];
     for (const [key, value] of Object.entries(MEASUREMENT_UNITS)) {
-        if (value.label === 'unit') options.push({ value: value.label, label: value.label });
-        else options.push({ value: value.label, label: value.label + ' (' + value.longLabelPlural + ')' });
+        if (value.label === 'unit') {
+            options.push({value: value.label, label: value.label});
+        } else {
+            options.push({value: value.label, label: value.label + ' (' + value.longLabelPlural + ')'});
+        }
     }
     return options;
 }
 
 export function getAltMetricUnitOptions(unitTypeStr: string, showLongLabel?: boolean): any[] {
     const unit: MeasurementUnit = MEASUREMENT_UNITS[unitTypeStr?.toLowerCase()];
-    if (!unit || unit.baseUnit === BASE_UNITS.COUNT) return [];
+    if (!unit || unit.baseUnit === BASE_UNITS.COUNT) {
+        return [];
+    }
 
     const options = [];
     for (const [key, value] of Object.entries(MEASUREMENT_UNITS)) {
         if (value.baseUnit === unit.baseUnit) {
-            if (!showLongLabel) options.push({ value: value.label, label: value.label });
-            else options.push({ value: value.label, label: value.label + ' (' + value.longLabelPlural + ')' });
+            if (!showLongLabel) {
+                options.push({value: value.label, label: value.label});
+            } else {
+                options.push({value: value.label, label: value.label + ' (' + value.longLabelPlural + ')'});
+            }
         }
     }
     return options;
@@ -99,11 +228,15 @@ export function getAltMetricUnitOptions(unitTypeStr: string, showLongLabel?: boo
 
 export function getAltUnitKeys(unitTypeStr): string[] {
     const unit: MeasurementUnit = MEASUREMENT_UNITS[unitTypeStr?.toLowerCase()];
-    if (!unit) return [];
+    if (!unit) {
+        return [];
+    }
 
     const options = [];
     Object.values(MEASUREMENT_UNITS).forEach(value => {
-        if (value.baseUnit === unit.baseUnit) options.push(value.label);
+        if (value.baseUnit === unit.baseUnit) {
+            options.push(value.label);
+        }
     });
 
     return options;
@@ -119,7 +252,9 @@ export function getMultiAltUnitKeys(unitTypeStrs: string[]): string[] {
         }
     }
 
-    if (!compabitable) return [];
+    if (!compabitable) {
+        return [];
+    }
 
     if (!unitTypeStr) {
         // if no unit type, return all unit types
@@ -152,7 +287,9 @@ export function convertUnitDisplay(
 
     const currentUnit: MeasurementUnit = MEASUREMENT_UNITS[unit.toLowerCase()];
     const targetUnit: MeasurementUnit = MEASUREMENT_UNITS[displayUnit.toLowerCase()];
-    if (!currentUnit || !targetUnit) return amount + (includeUnits ? ' ' + unit : '');
+    if (!currentUnit || !targetUnit) {
+        return amount + (includeUnits ? ' ' + unit : '');
+    }
 
     const newAmount = amount * (currentUnit.ratio / targetUnit.ratio);
     // show up to 6 decimal points
@@ -165,17 +302,23 @@ export function convertUnitDisplay(
 // volume: 10
 // volume not set: null
 export function getStoredAmountDisplay(rawValue: string, includeUnits?: boolean): string {
-    if (!rawValue || !rawValue.trim()) return null;
+    if (!rawValue || !rawValue.trim()) {
+        return null;
+    }
 
     const parts: string[] = rawValue.trim().split(/\b\s+/);
-    if (parts.length === 1) return rawValue;
+    if (parts.length === 1) {
+        return rawValue;
+    }
 
     if (parts.length === 2) {
         if (includeUnits) {
             const unit = parts[1];
-            if (unit.indexOf('(') === 0 && unit.indexOf(')') === unit.length - 1)
+            if (unit.indexOf('(') === 0 && unit.indexOf(')') === unit.length - 1) {
                 return parts[0] + ' ' + unit.substring(1, unit.length - 1);
-            else return rawValue;
+            } else {
+                return rawValue;
+            }
         } else {
             return parts[0];
         }
@@ -198,18 +341,22 @@ export function getStoredAmountDisplay(rawValue: string, includeUnits?: boolean)
 
 export function getVolumeMinStep(sampleTypeUnit?: string | MeasurementUnit) {
     const step = 0.01;
-    if (!sampleTypeUnit) return step;
+    if (!sampleTypeUnit) {
+        return step;
+    }
 
     const unit = typeof sampleTypeUnit === 'string' ? MEASUREMENT_UNITS[sampleTypeUnit.toLowerCase()] : sampleTypeUnit;
 
     // If we don't know the units, or it is 'unit' then use the default
-    if (!unit || unit === MEASUREMENT_UNITS.unit) return step;
+    if (!unit || unit === MEASUREMENT_UNITS.unit) {
+        return step;
+    }
 
     return Math.pow(10, -unit.displayPrecision); // Track uL and mg to a single unit
 }
 
 export function isMeasurementUnitIgnoreCase(expected: MeasurementUnit, val: string) {
-    return expected.label.localeCompare(val, 'en-US', { sensitivity: 'base' }) === 0;
+    return expected.label.localeCompare(val, 'en-US', {sensitivity: 'base'}) === 0;
 }
 
 export function isValuePrecisionValid(value: number, precision: number): boolean {
